backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Anthropic Claude API
anthropic==0.18.1

# DOCX processing
python-docx==1.1.0
lxml==5.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# TMX and fuzzy matching
rapidfuzz==3.6.1
lxml==5.1.0

# HTTP clients for HUDOC/CURIA
httpx==0.26.0
beautifulsoup4==4.12.3

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
